Separate the adapter from the activity
